feat: Create main layout areas for logged in pages
Adjust the layout of the application to use a two-column approach once
the user is logged in.
A logged out user will still see the centered one-column layout that
will cater to the landing page and authentication use cases.

This commit also reorganizes the theme related code into #app/theme
directory.