feat: Add support for theme colors in tailwind.css
Theme colors in Tailwind are defined using HEX codes. Overall, the
theming in the epic-stack is setup to use HSL, leading to some hardcoded
values for colors in tailwind.css file.

By adding 'postcss-color-converter' plugin, we can now use
theme(<tailwind-color>) in tailwind.css file, which will be converted to
HSL color space.

Because of this, we need to remove the explicit hsl() call in
extended-theme.ts as the variables will already have hsl().